# Conflicts:
#	Robust.Client/GameObjects/Components/Light/PointLightComponent.cs
#	Robust.Client/Graphics/Clyde/Clyde.LightRendering.cs
#	Robust.Client/Graphics/Clyde/Clyde.cs
#	Robust.Client/ResourceManagement/ResourceCache.cs

+++ Robust.Client/Map/IClydeTileDefinitionManager.cs
@@ -1,5 +1,6 @@
 using System.Collections.Generic;
 using Robust.Client.Graphics;
+using Robust.Shared.Graphics;
 using Robust.Shared.Map;
 using Robust.Shared.Maths;
 
@@ -27,6 +28,6 @@ namespace Robust.Client.Map
         ///     Gets the region inside the texture atlas to use to draw a tile type.
         /// </summary>
         /// <returns>If null, do not draw the tile at all.</returns>
-        Box2[]? TileAtlasRegion(ushort tileType);
+        Box2[]? TileAtlasRegion(int tileType);
     }
 }
